webserver/pyproject.toml
[project]
name = "webserver"
version = "0.1.0"
description = "reverse proxy games"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "elasticsearch>=9.0.1",
    "fastapi>=0.115.12",
    "ipwhois>=1.3.0",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

